requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
tiktoken==0.5.2
sentence-transformers==2.2.2
